Mockup of new mode selection UI
Major feedback from day 1-3 of the arcade that's hosting the Simply Spud cab's opening: SL's mode selection UI works great for experienced players, but confuses the heck out of casual and mid-tier players. Given that this theme targets public cabs (namely ours), this should be addressed in this fork, and hopefully this commit is a step towards that.

(Layout suggestion by Tiktalik with feedback from the rest of the group)

 - ScreenSelectPlayMode now has two icons - One for casual, one for standard mode (FA+ support will likely be readded in another form, later)
 - New *PLACEHOLDER* graphics icons for ScreenSelectPlayMode and ScreenSelectPlayMode2
 - Mode name changes: Casual -> Beginner, ITG -> Pro (hopefully makes it clearer to casual players that they shouldn't pick ITG/FA+/etc mode)
 - Mode descriptions are now shorter (might be tweaked again in the future, wordiness probably isn't great for casual players either)
 - ScreenSelectPlayMode now has some slightly fancier menu transitions
 - Tweaked the title screen flying <xyz objects> to fade out a bit more smoothly
 - Removed the now unused gameplay preview from ScreenSelectPlayMode