Initial ITGMania/GrooveStats Integration
Thanks to SteveReen and teejusb for giving some pointers for getting started with this! It was suggested to copy many of these bits from mainline SL, so credit for most of this goes to the mainline SL team.

This commit uses ITGMania's network calls for a WIP GrooveStats integration, thus ITGMania is now required. Only score loading on the music wheel is implemented at this time, but other functionality will likely be implemented in the future!

 - Add ThemePref, Helpers, Profile Load, and SystemLayer functionality for GrooveStats from SL
 - Update chart parsing/hashing, SL QR code generation, timing window options, and emojis to SL latest
 - Rework DetailsDisplay to show rival and world high scores
 - Fix ThemePrefs validation reading from Simply Love (and not Spud)
 - Fix the default SL color (we're Orange now, not Purple)